Fixed bug drawing player health bar - ints should be cast to floats before mult/div ops.

+++ core/src/main/java/com/genericgames/samurai/screens/GameView.java
@@ -346,7 +346,7 @@ public class GameView extends StageView {
         Rectangle healthBarRect = healthBar.getRectangle();
         shapeRenderer.rect(healthBarRect.getX(), healthBarRect.getY(),
                 healthBarRect.getWidth(),
-                healthBarRect.getHeight()*(samuraiWorld.getPlayerCharacter().getHealth()/samuraiWorld.getPlayerCharacter().getMaxHealth()));
+                healthBarRect.getHeight()*((float)samuraiWorld.getPlayerCharacter().getHealth()/(float)samuraiWorld.getPlayerCharacter().getMaxHealth()));
         shapeRenderer.end();
 
         shapeRenderer.begin(ShapeRenderer.ShapeType.Line);
